Add FMS API OpenAI and SignalR Hubs
We now have an OpenAPI file representing FMS APIs, and this is used to generate an API implementation for us. It also has information about the SignalR hubs we use for real-time communication, but as those don't document very well with swagger/OpenAPI, I've included documentation on the hubs in markdown and implemented a simple API for starting the hubs and registering callbacks for the hub events.